+++ app/src/main/java/com/techcourse/dao/JdbcUserHistoryDao.java
@@ -1,0 +1,68 @@
+package com.techcourse.dao;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+import javax.sql.DataSource;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.techcourse.domain.UserHistory;
+
+import nextstep.jdbc.JdbcTemplate;
+
+public class JdbcUserHistoryDao implements UserHistoryDao {
+
+    private static final Logger log = LoggerFactory.getLogger(JdbcUserHistoryDao.class);
+
+    private final DataSource dataSource;
+
+    public JdbcUserHistoryDao(final DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
+
+    public JdbcUserHistoryDao(final JdbcTemplate jdbcTemplate) {
+        this.dataSource = null;
+    }
+
+    @Override
+    public void log(final UserHistory userHistory) {
+        final var sql = "insert into user_history (user_id, account, password, email, created_at, created_by) values (?, ?, ?, ?, ?, ?)";
+
+        Connection conn = null;
+        PreparedStatement pstmt = null;
+        try {
+            conn = dataSource.getConnection();
+            pstmt = conn.prepareStatement(sql);
+
+            log.debug("query : {}", sql);
+
+            pstmt.setLong(1, userHistory.getUserId());
+            pstmt.setString(2, userHistory.getAccount());
+            pstmt.setString(3, userHistory.getPassword());
+            pstmt.setString(4, userHistory.getEmail());
+            pstmt.setObject(5, userHistory.getCreatedAt());
+            pstmt.setString(6, userHistory.getCreateBy());
+            pstmt.executeUpdate();
+        } catch (SQLException e) {
+            log.error(e.getMessage(), e);
+            throw new RuntimeException(e);
+        } finally {
+            try {
+                if (pstmt != null) {
+                    pstmt.close();
+                }
+            } catch (SQLException ignored) {
+            }
+
+            try {
+                if (conn != null) {
+                    conn.close();
+                }
+            } catch (SQLException ignored) {
+            }
+        }
+    }
+}
