+++ app/src/test/java/com/techcourse/dao/JdbcUserDaoTest.java
@@ -9,15 +9,15 @@ import com.techcourse.config.DataSourceConfig;
 import com.techcourse.domain.User;
 import com.techcourse.support.jdbc.init.DatabasePopulatorUtils;
 
-class UserDaoTest {
+class JdbcUserDaoTest {
 
-    private UserDao userDao;
+    private JdbcUserDao userDao;
 
     @BeforeEach
     void setup() {
         DatabasePopulatorUtils.execute(DataSourceConfig.getInstance());
 
-        userDao = new UserDao(DataSourceConfig.getInstance());
+        userDao = new JdbcUserDao(DataSourceConfig.getInstance());
         final var user = new User("gugu", "password", "hkkang@woowahan.com");
         userDao.save(user);
     }
